requirements.txt
torch>=2.0.0
sentencepiece
faiss-cpu
bitsandbytes
pyyaml
tqdm
numpy
